+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.35.1-0 / Unreleased
   * [Fixed] #1001 Invalid urls in skin-xp CSS
+  * [Changed] #1005 Cast key to string in getNodeByKey()
 
 # 2.35.0 / 2020-03-27
   * [Changed] The `enableAspx` option will default to 'false' in the future.